backend/requirements.txt
aiobotocore==2.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
amqp==5.3.1
aniso8601==10.0.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.0.1
billiard==3.6.4.0
boto3==1.26.76
botocore==1.29.76
celery==5.2.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.7.0
email-validator==1.3.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
frozenlist==1.5.0
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.0
MarkupSafe==3.0.2
marshmallow==3.19.0
multidict==6.2.0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2025.1
redis==4.5.1
requests==2.28.2
s3transfer==0.6.2
setuptools==76.1.0
six==1.17.0
structlog==23.1.0
tzdata==2025.1
urllib3==1.26.20
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.17.2
yarl==1.18.3
